requirements.txt
aiohttp==3.8.1
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cmake==3.21.1.post1
cryptography==36.0.2
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
et-xmlfile==1.1.0
facebook-sdk==3.1.0
filelock==3.6.0
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-Table==0.5.0
Flask-User==1.0.2.2
Flask-WTF==1.0.0
folium==0.12.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
guizero==1.3.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.1
MarkupSafe==2.1.1
mongoengine==0.24.1
multidict==6.0.2
numpy==1.21.5
oauthlib==3.2.0
openpyxl==3.0.9
outcome==1.1.0
passlib==1.7.4
Pillow==9.0.1
praw==7.5.0
prawcore==2.3.0
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pymongo==4.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.8.0
typing-extensions==4.1.1
update-checker==0.18.0
urllib3
virtualenv==20.14.0
websocket-client==1.3.1
Werkzeug==2.0.3
wsproto==1.1.0
WTForms==3.0.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.7.0
